agents/linkedin-agent/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
python-dotenv>=1.0
requests>=2.31
firebase-admin>=6.5.0
